+++ Tests/Integration/ZendLuceneAdapterTest.php
@@ -12,6 +12,7 @@ namespace Massive\Bundle\SearchBundle\Tests\Integration;
 
 use Symfony\Component\Filesystem\Filesystem;
 use Massive\Bundle\SearchBundle\Search\Adapter\ZendLuceneAdapter;
+use Massive\Bundle\SearchBundle\Search\Localization\NoopStrategy;
 
 class ZendLuceneAdapterTest extends AdapterTestCase
 {
@@ -20,13 +21,17 @@ class ZendLuceneAdapterTest extends AdapterTestCase
 
     public function setUp()
     {
-        parent::setUp();
-        $this->baseDir = sys_get_temp_dir() . '/massive-test-zend-lucene';
         $this->filesystem = new Filesystem();
+        $this->baseDir = sys_get_temp_dir() . '/massive-test-zend-lucene';
+        parent::setUp();
+    }
 
-        if (!file_exists($this->baseDir)) {
-            mkdir($this->baseDir);
+    public function purgeIndex($indexName)
+    {
+        if (file_exists($this->baseDir)) {
+            $this->filesystem->remove($this->baseDir);
         }
+        $this->filesystem->mkdir($this->baseDir);
     }
 
     public function tearDown()
@@ -34,8 +39,21 @@ class ZendLuceneAdapterTest extends AdapterTestCase
         $this->filesystem->remove($this->baseDir);
     }
 
-    public function getAdapter()
+    public function doGetAdapter()
     {
-        return new ZendLuceneAdapter($this->getFactory(), $this->baseDir);
+        return new ZendLuceneAdapter($this->getFactory(), new NoopStrategy(), $this->baseDir, true);
+    }
+
+    public function provideSearch()
+    {
+        return array(
+            array('one', 1),
+            array('one ', 1),
+            array('roomba 870', 0),
+            array('870', 0),
+            array('*', 0),
+            array('***', 0),
+            array('???', 0),
+        );
     }
 }